requirements.txt
asgiref==3.8.1
cloudinary==1.39.1
dj-database-url==2.1.0
Django==5.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
PyJWT==2.8.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1